add Friends, likes, exception

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -2,13 +2,17 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDate;
+import java.util.HashSet;
+import java.util.Set;
 
 @Data
 @AllArgsConstructor
 public class Film {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private LocalDate releaseDate;
